lemon_agent_implementation/requirements.txt
openai>=1.0.0
google-genai
google-cloud-aiplatform
